feat(cython): move safe tensor surface into TensorImpl base (#164)
Move read-only metadata, scalar protocol, dtype shortcuts, comparison
dunders, and representation methods from the Python Tensor class into
the existing Cython TensorImpl base (and its pure-Python fallback).

Migrated surface (41 methods/properties):
- Read-only metadata: output_nr, is_cuda, is_cpu, is_npu, is_meta,
  is_leaf, is_sparse, layout, is_quantized, storage_offset, get_device,
  ndimension, size, nelement
- Scalar protocol: item, tolist, __int__, __float__, __bool__
- Representation: __repr__, __str__, __len__, __iter__, __hash__
- Dtype wrappers: float, half, double, bfloat16, long, int, short,
  char, byte, bool
- Comparison dunders: __gt__, __lt__, __ge__, __le__, __eq__, __ne__
- Helper: _is_scalar_comparable

No changes to autograd, view, inplace, or version-counter semantics.
No new extension target or build-system complexity.
